requirements.txt
fastapi==0.115.11
playwright==1.40.0
pydantic==2.10.6
pymongo==4.11.2
Requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
